+++ paddle/fluid/operators/lstm_cudnn_op.cu.cc
@@ -12,7 +12,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-#include "paddle/fluid/operators/cudnn_lstm_op.h"
+#include "paddle/fluid/operators/lstm_cudnn_op.h"
 #include "paddle/fluid/platform/cudnn_helper.h"
 
 namespace paddle {
